refactor(zenhub): move execution to new directory path
